requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
asttokens==3.0.1
attrs==25.4.0
Authlib==1.6.5
autoflake==2.3.1
beautifulsoup4==4.14.2
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
comm==0.2.3
contourpy==1.3.3
cryptography==46.0.3
cycler==0.12.1
debugpy==1.8.17
decorator==5.2.1
distro==1.9.0
dparse==0.6.4
executing==2.2.1
fastapi==0.122.0
filelock==3.20.0
flake8==7.3.0
fonttools==4.61.0
frozenlist==1.8.0
googleapis-common-protos==1.72.0
greenlet==3.2.4
grimp==3.13
grpcio==1.76.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
import-linter==2.7
importlib_metadata==8.7.0
iniconfig==2.3.0
ipykernel==7.1.0
ipython==9.7.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jiter==0.12.0
joblib==1.5.2
jupyter_client==8.6.3
jupyter_core==5.9.1
kiwisolver==1.4.9
librt==0.6.3
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==4.1.0
matplotlib==3.10.7
matplotlib-inline==0.2.1
mccabe==0.7.0
mdurl==0.1.2
mpmath==1.3.0
multidict==6.7.0
mypy==1.19.0
mypy_extensions==1.1.0
nest-asyncio==1.6.0
nltk==3.9.2
numpy==2.2.6
openai==2.8.1
opencv-python==4.12.0.88
opentelemetry-api==1.38.0
opentelemetry-exporter-otlp==1.38.0
opentelemetry-exporter-otlp-proto-common==1.38.0
opentelemetry-exporter-otlp-proto-grpc==1.38.0
opentelemetry-exporter-otlp-proto-http==1.38.0
opentelemetry-proto==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
packaging==25.0
parso==0.8.5
pathspec==0.12.1
pexpect==4.9.0
pillow==12.0.0
platformdirs==4.5.0
playwright==1.56.0
plt==0.2.0
pluggy==1.6.0
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.1
psutil==7.1.3
ptyprocess==0.7.0
pure_eval==0.2.3
pycodestyle==2.14.0
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pyee==13.0.0
pyflakes==3.4.0
Pygments==2.19.2
pyparsing==3.2.5
pytesseract==0.3.13
pytest==9.0.1
python-dateutil==2.9.0.post0
python-multipart==0.0.20
PyYAML==6.0.3
pyzmq==27.1.0
regex==2025.11.3
requests==2.32.5
rich==14.2.0
ruamel.yaml==0.18.16
safety==3.7.0
safety-schemas==0.0.16
scikit-learn==1.7.2
scipy==1.16.3
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
stack-data==0.6.3
starlette==0.50.0
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tomlkit==0.13.3
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
# Editable Git install with no remote (troostwatch==0.7.0)
-e /workspaces/Troostwatch
typer==0.20.0
types-requests==2.32.4.20250913
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0
wcwidth==0.2.14
websockets==15.0.1
yarl==1.22.0
zipp==3.23.0
